git-svn-id: file:///srv/local/data/cleanRepo/trunk@8063 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/playground/jjoubert/CommercialDemand/MyGrid.java
@@ -151,7 +151,7 @@ public class MyGrid {
 				int hour = Integer.parseInt(line[3].substring(position-2, position));
 				
 				Collection<MyGridCell> cells = grid.get(x, y, radius);
-				double value = 1.0 / (double) cells.size();
+				float value = ((float) 1) / ((float) cells.size());
 				for (MyGridCell cell : cells) {
 					cell.addToTotalCount(value);
 					cell.addToHourCount(hour, value);
